Refactor project structure: move scripts to a dedicated folder, implement WebSocket handling, and enhance drag-and-drop functionality for player/spectator management.
