+++ front-end/src/pages/index.astro
@@ -140,7 +140,7 @@ let showJoinRole = true
     </div>
 </Main>
 
-<script type="module" src="/src/game.ts"></script>
+<script type="module" src="/src/scripts/game.ts"></script>
 
 <style>
     #gameCode {
