+++ front-end/src/scripts/game.ts
